satagro/requirements.txt
Django==4.0.1
psycopg2>=2.8
uuid
djangorestframework
requests
